chore: ProductItem 통일
